Update manpage.
In addition to the various obvious changes we need to make with the
switch to libgweather, we also change how the manpage is generated.
In particular, because we want to substitue in $pkgdatadir variable
(for documenting icondir), we manually call sed in the Makefile instead
of using AC_CONFIG_FILES, as outlined in section 9.2.3 of the
autotools manual.