debian/source/options: Remove file; not necessary.
I added this (and the upstream ChangeLog to git) because I didn't understand
the bug that prompted the bump to version 1.2.  It wasn't that that
we needed ChangeLog to be in the source directory -- it already was because
it's in the orig tarball!  It's that we needed to be able to find it
when building from another, which now happens.